+++ spec/buildpack/detect_spec.rb
@@ -24,29 +24,14 @@ describe AspNet5Buildpack::Detecter do
   end
 
   describe 'detect' do
-    context 'when there is no NuGet.Config in the root directory' do
-      before do
-        File.open(File.join(dir, 'project.json'), 'w') { |f| f.write('a') }
-      end
-
-      it 'returns false' do
-        expect(subject.detect(dir)).to be_falsey
-      end
-    end
-
     context 'when there is no project.json' do
-      before do
-        File.open(File.join(dir, 'NuGet.Config'), 'w') { |f| f.write('a') }
-      end
-
       it 'returns false' do
         expect(subject.detect(dir)).to be_falsey
       end
     end
 
-    context 'when project.json and NuGet.Config exist in the same directory' do
+    context 'when project.json exists in the root directory' do
       before do
-        File.open(File.join(dir, 'NuGet.Config'), 'w') { |f| f.write('a') }
         File.open(File.join(dir, 'project.json'), 'w') { |f| f.write('a') }
       end
 
@@ -55,9 +40,8 @@ describe AspNet5Buildpack::Detecter do
       end
     end
 
-    context 'when project.json and NuGet.Config exist in different directories' do
+    context 'when project.json exists in a sub-directory' do
       before do
-        File.open(File.join(dir, 'NuGet.Config'), 'w') { |f| f.write('a') }
         FileUtils.mkdir_p(File.join(dir, 'src', 'proj'))
         File.open(File.join(dir, 'src', 'proj', 'project.json'), 'w') { |f| f.write('a') }
       end