+++ lib/buildpack/compile/dnu.rb
@@ -24,10 +24,9 @@ module AspNet5Buildpack
 
     def restore(dir, out)
       @shell.env['HOME'] = dir
-      @shell.env['MONO_THREADS_PER_CPU'] = '2000'
-      @shell.path << '/app/mono/bin'
+      @shell.env['LD_LIBRARY_PATH'] = "$LD_LIBRARY_PATH:#{dir}/libunwind/lib"
       project_list = AppDir.new(dir, out).with_project_json.join(' ')
-      cmd = "bash -c 'source #{dir}/.dnx/dnvm/dnvm.sh; dnvm use default -r mono -a x64; cd #{dir}; dnu restore #{project_list}'"
+      cmd = "bash -c 'source #{dir}/.dnx/dnvm/dnvm.sh; dnvm use default; cd #{dir}; dnu restore #{project_list}'"
       @shell.exec(cmd, out)
     end
   end